python3: version bump 3.5.4 -> 3.7.0
zlib support can no longer be disabled. 'crc32' from zlib is used in other
modules if found. So not installing the zlib module is no longer sufficient
to ensure that zlib is not used. zlib is small enough so just depend on it
unconditionally.

Also, setup the environment to make the build reproducible.

Depend on host-python3-setuptools. This is needed by most packages anyways,
so we can drop a lot of extra dependencies. Also, we need to import some of
it once: It generates some pickle files and that's racy during parallel
building. These files should be generated at build time, but that happens
in pkgdir and for cross we use the source tree in build-target.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>