ptxd_make_world_init: reset SOURCE_DATE_EPOCH for local-src packages
If a temporary symlink in local_src/ is used to provide the source code for
a package, then reset SOURCE_DATE_EPOCH to the current date/time.

Such symlinks are used during development only and built-in timestamps can
be used to determine that the correct version of a binary is used.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>