openssl: remove engines-path patch
It is a Debian specific patch that changes the engines path to a version
specific directory. This is not needed here because we only have one
openssl version.
The openssl version is needed to predict this path. This can be difficult
to do outside the openssl package.
Just remove the patch and use the default path.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>