rules: allow selecting most host packages in the platformconfig
Do this by including the the hosttools_noprompt section for the
platformconfig.
This way, we don't need to manually add a package to the platformconfig
when it is needed. Exceptions are special packages that are always enabled
and host-gobject-introspection because it needs host-qemu. These packages
are now in the new hosttools_noprompt_ptx section.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>