ptxd_init_*_env: use -B instead of -L
It's not quite correct, but it shouldn't hurt and with -B the path is added
to 'libraries:' in 'gcc -print-search-dirs'. Libtool uses this for manual
seaching. This avoid explicit paths to .la files found in the system when
building host packages.
Also assume -print-search-dirs means compiling/linking to ensure the
options are used.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>