ptxd_kconfig: fix restoring missing diff
If the diff is missing, then it must be regenerated from base and current
config. If the current config is up to date then no config files are
written and no diff is generated.
Force a config update if the diff is missing to ensure a new one is
generated.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>