ptxd_lib_dgen: fix pkghash for manually included menu files
Manual includes are relative to the temporary kconfig file tree. The
current working directory for ptxd_lib_dgen.awk is the workspace, so these
relative files are not found.
Add the correct absolute prefix for these files to fix this.
Also, make 'tmp' und dump_file() a local variable to ensure that it does
not contain old contents.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>