template-src-linux-driver: Only install driver if selected
If a local package "foo-mod" for a Linux kernel module is selected by a
local tristate configuration option PTXCONF_FOO_MOD, then based on the
template for Linux drivers, it would include the following in its
foo-mod.make file:

ifdef PTXCONF_FOO_MOD
$(STATEDIR)/kernel.targetinstall.post: $(STATEDIR)/foo-mod.targetinstall
endif

However, this rule will be invoked even if the PTXCONF_FOO_MOD option
has been configured with the tristate option 'm' and has not been
changed to 'y' by a selected_collectionconfig.  This causes the Linux
kernel module to be built and installed unintentionally.  To prevent
that happening, the above lines in the foo-mod.make file can be changed
as follows (using a conditional test suggested by Michael Olbrich):

ifneq ($(filter foo-mod,$(PTX_PACKAGES_SELECTED)),)
$(STATEDIR)/kernel.targetinstall.post: $(STATEDIR)/foo-mod.targetinstall
endif

This patch changes the template for Linux driver packages in accordance
with the above change.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>
Cc: Michael Olbrich <m.olbrich@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>