libpaper: Make the default paper configurable
For embedded systems it makes sense to be able to configure the default
paper name at build-time instead of run-time via environment variable.

The approach currently used in the rule file to configure the paper size
at build-time does not work, since using the environment variable
'PAPERSIZE' only works at run-time.

This change adds a new '--with-default-paper' configure switch, to be able to
overwrite the default 'letter' setting.

Signed-off-by: Juergen Borleis <jbe@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>