ptxd_install_glob: fix glob argument splitting
The shell is used to split the strings into multiple array elements.
However, this means that other expansions are performed as well. E.g. path
expansion (with workspace as current directory). So files that match a
pattern break the build.
Disable pathname and brace expansion temporarily to fix this. Use an array
afterwards to ensure that the patterns can be quoted correctly afterwards.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>